OnClick handlers moved to correct controllers.
