+++ Frontend/src/main/java/com/harmonia/controller/ChatController.java
@@ -1,20 +1,113 @@
 package com.harmonia.controller;
 
+import java.io.IOException;
 import java.util.*;
 
+import com.harmonia.HarmoniaApplication;
 import com.harmonia.model.Message;
 import com.harmonia.view.ChatView;
+import javafx.event.ActionEvent;
+import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
+import javafx.stage.Stage;
 
 public class ChatController {
     private ChatView view;
     private List<Message> messages;
+
+    /**
+     navigation button for nav menu
+     letter combination before name indicates in what view the button is from
+     h=harmonia-view
+     fm=messages-view
+     p=profile-view
+     */
+    @FXML
+    private Button fmFriendsBtn;
+
+    /**
+     navigation button for nav menu
+     letter combination before name indicates in what view the button is from
+     h=harmonia-view
+     fm=messages-view
+     p=profile-view
+     */
+    @FXML
+    private Button fmSettingsBtn;
+
+    /**
+     navigation button for nav menu
+     letter combination before name indicates in what view the button is from
+     h=harmonia-view
+     fm=messages-view
+     p=profile-view
+     */
+    @FXML
+    private Button fmProfileBtn;
+
+    /**
+     navigation button for nav menu
+     letter combination before name indicates in what view the button is from
+     h=harmonia-view
+     fm=messages-view
+     p=profile-view
+     */
+    @FXML
+    private Button fmHomePageBtn;
+
+    /**
+     navigation button for nav menu
+     letter combination before name indicates in what view the button is from
+     h=harmonia-view
+     fm=messages-view
+     p=profile-view
+     */
+    @FXML
+    private Button fmCommunityBtn;
 
     public ChatController(ChatView view) {
         this.view = view;
         this.messages = new ArrayList<>();
     }
 
+    public ChatController(){
+
+    }
+
     public void sendMessage(Message message) {
         // Add the message to the list and update the view
+    }
+
+    @FXML
+    protected void onfmHomePageBtnClick(ActionEvent event) {
+        try {
+            FXMLLoader fxmlLoader = new FXMLLoader(HarmoniaApplication.class.getResource("harmonia-view.fxml"));
+            Parent root = fxmlLoader.load();
+            Scene scene = new Scene((root), 1280, 720);
+            Stage stage = (Stage) fmHomePageBtn.getScene().getWindow();
+            stage.setScene(scene);
+            stage.setTitle("Sign in to Harmonia");
+            stage.show();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    @FXML
+    protected void onfmProfileBtnClick(ActionEvent event) {
+        try {
+            FXMLLoader fxmlLoader = new FXMLLoader(HarmoniaApplication.class.getResource("profile-view.fxml"));
+            Parent root = fxmlLoader.load();
+            Scene scene = new Scene((root), 1280, 720);
+            Stage stage = (Stage) fmProfileBtn.getScene().getWindow();
+            stage.setScene(scene);
+            stage.setTitle("Sign in to Harmonia");
+            stage.show();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
     }
 }
