+++ Frontend/src/main/java/com/harmonia/controller/LoginController.java
@@ -3,9 +3,11 @@ package com.harmonia.controller;
 import com.harmonia.HarmoniaApplication;
 import com.harmonia.client.UserClient;
 
+import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Node;
+import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.control.Hyperlink;
@@ -13,6 +15,8 @@ import javafx.scene.control.Label;
 import javafx.scene.control.PasswordField;
 import javafx.scene.control.TextField;
 import javafx.stage.Stage;
+
+import java.io.IOException;
 
 public class LoginController {
 
@@ -30,9 +34,6 @@ public class LoginController {
 
     @FXML
     Label errorLabel;
-
-
-    
 
     @FXML
     public void onLoginButtonClick() {
